fix(arc-740): make generic visit request table for cp and meemoo admins
