+++ src/modules/cp/components/index.ts
@@ -1,7 +1,2 @@
-export * from './RequestStatusBadge';
-export * from './ApproveRequestBlade';
-export * from './DeclineRequestBlade';
-export * from './ProcessRequestBlade';
 export * from './ReadingRoomImageForm';
 export * from './ReadingRoomSettings';
-export * from './RichTextForm';
