+++ src/modules/cp/const/visitors.const.tsx
@@ -3,14 +3,13 @@ import { isWithinInterval } from 'date-fns';
 import { Column } from 'react-table';
 import { NumberParam, StringParam, withDefault } from 'use-query-params';
 
-import { RequestStatusAll } from '@cp/types';
 import { CopyButton, DropdownMenu, UnreadMarker } from '@shared/components';
 import { SEARCH_QUERY_KEY } from '@shared/const';
 import { SortDirectionParam } from '@shared/helpers';
 import { i18n } from '@shared/helpers/i18n';
 import { OrderDirection, Visit, VisitRow } from '@shared/types';
 import { asDate, formatSameDayRange } from '@shared/utils';
-import { VisitTimeframe } from '@visits/types';
+import { RequestStatusAll, VisitTimeframe } from '@visits/types';
 
 export const CP_ADMIN_VISITORS_QUERY_PARAM_CONFIG = {
 	timeframe: withDefault(StringParam, RequestStatusAll.ALL),
