--- src/modules/cp/types/index.ts
@@ -1,1 +1,0 @@
-export * from './requests.types';
