+++ src/modules/shared/components/DeclineRequestBlade/DeclineRequestBlade.types.ts
