+++ src/modules/shared/components/index.ts
@@ -1,3 +1,4 @@
+export * from './ApproveRequestBlade';
 export * from './Blade';
 export * from './BladeManager';
 export * from './Callout';
@@ -5,6 +6,7 @@ export * from './CardImage';
 export * from './ConfirmationModal';
 export * from './CopyButton';
 export * from './Datepicker';
+export * from './DeclineRequestBlade';
 export * from './DropdownMenu';
 export * from './ErrorNotFound';
 export * from './ErrorPage';
@@ -20,8 +22,11 @@ export * from './Overlay';
 export * from './PaginationBar';
 export * from './PaginationProgress';
 export * from './Placeholder';
+export * from './ProcessRequestBlade';
 export * from './ReadingRoomCard';
 export * from './ReadingRoomCardList';
+export * from './RequestStatusBadge';
+export * from './RichTextEditor';
 export * from './SearchBar';
 export * from './Sidebar';
 export * from './SidebarLayoutTitle';
