+++ src/modules/visits/components/VisitRequestsOverview/VisitRequestsOverview.types.tsx
@@ -1,0 +1,7 @@
+import { Column } from 'react-table';
+
+import { Visit } from '@shared/types';
+
+export interface VisitRequestOverviewProps {
+	columns: Column<Visit>[];
+}
